projects/policy-chatbot/src/pyproject.toml
[project]
name = "policy-chatbot"
version = "0.1.0"
description = "Corporate Policy Assistant Chatbot — RAG-powered conversational AI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "redis>=5.2.0",
    "azure-identity>=1.19.0",
    "azure-storage-blob>=12.24.0",
    "azure-search-documents>=11.6.0",
    "openai>=1.58.0",
    "azure-monitor-opentelemetry>=1.6.4",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "python-multipart>=0.0.18",
    "httpx>=0.28.0",
    "python-jose[cryptography]>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "W",    # pycodestyle warnings
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "S",    # flake8-bandit (security)
    "A",    # flake8-builtins
    "C4",   # flake8-comprehensions
    "PT",   # flake8-pytest-style
    "RUF",  # ruff-specific rules
    "T20",  # flake8-print (no print statements)
]

[tool.ruff.lint.per-file-ignores]
"alembic/*" = ["E501", "I001"]
"tests/*" = ["S101"]
"app/api/*" = ["B008"]
"app/core/auth.py" = ["B008"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
exclude = ["alembic/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["../tests"]
